requirements.txt
tensorflow==1.13.1
opencv-python==3.4.2.17
imutils==0.5.3
scikit-image==0.16.1
pyodbc==4.0.30
PyQt5==5.15.2
PySide2==5.15.2
pyserial==3.5
configparser==5.0.1